meta-qti-bsp: Update unifdef path only if its exported

+++ recipes-kernel/linux-msm/linux-soc-repo_6.%.bb
@@ -92,9 +92,11 @@ do_prebuilt_shared_workdir() {
         install -m 0644 ${B}/scripts/module.lds ${STAGING_KERNEL_BUILDDIR}/scripts/module.lds
     fi
 
-    cp -rp ${KERNEL_PREBUILT_PATH}/host/unifdef $kerneldir/scripts/
-    cp -rp ${KERNEL_PLATFORM_PATH}/soc-repo/scripts/headers_install.sh $kerneldir/
-    sed -i 's|scripts/unifdef|${STAGING_KERNEL_BUILDDIR}/scripts/unifdef|g' ${STAGING_KERNEL_BUILDDIR}/headers_install.sh
+    if [ -e "${KERNEL_PREBUILT_PATH}/host/unifdef" ]; then
+        cp -rp ${KERNEL_PREBUILT_PATH}/host/unifdef $kerneldir/scripts/
+        cp -rp ${KERNEL_PLATFORM_PATH}/soc-repo/scripts/headers_install.sh $kerneldir/
+        sed -i 's|scripts/unifdef|${STAGING_KERNEL_BUILDDIR}/scripts/unifdef|g' ${STAGING_KERNEL_BUILDDIR}/headers_install.sh
+    fi
 }
 
 do_prebuilt_install[dirs] = "${B}"
